test(cli): fix flaky dev-server HMR integration test
The test failed roughly 12% of runs with "timeout waiting for message
after 4000ms". Instrumenting which of the two waits expired showed the
flake was always on the `connected` ack and never on the hmr-update, so
the file watcher was never at fault — it delivered 30/30 in isolation.

ws emits 'message' on an EventEmitter, which does not buffer. The test
attached its listener only after `await waitForOpen(client)`. When the
handshake response and the ack arrive in the same client-side read, ws
emits 'open' and then 'message' inside one macrotask, before the awaited
continuation runs to subscribe. The ack is dropped and the test waits out
the full timeout for a message that already arrived.

hmr-server.js defers the ack with setTimeout(0) for this reason, but that
only narrows the window: the server cannot know when the client
subscribes. hmr-server.test.js had already worked around it with a local
bufferMessages helper; integration.test.js never adopted it.

Move bufferMessages and waitForOpen into a shared ws-helpers.js, subscribe
at socket construction in both suites, and report the message types
actually received when a wait does time out. 40/40 clean runs of the
previously flaky test and 25/25 of the whole dev-server suite.
